Use countries and accumulated repo in initializer
