+++ app/src/main/java/com/mobilecoronatracker/data/repository/AccumulatedDataRepo.kt
@@ -8,7 +8,8 @@ interface AccumulatedDataRepo {
     fun getTodayData(): Flow<GeneralReportModelable>
     fun getFullHistory(): Flow<List<GeneralReportTimePointModelable>>
     fun getDataForLastDays(numberOfDays: Int): Flow<List<GeneralReportModelable>>
-    suspend fun refreshAccumulatedData()
+    suspend fun fetchTodayAccumulatedData()
     suspend fun hasTodayData(): Boolean
     suspend fun hasAnyData(): Boolean
+    suspend fun fetchHistoricalAccumulatedData()
 }
